requirements.txt
streamlit==1.32.0
langchain==0.1.9
langchain-openai==0.0.8
python-dotenv==1.0.0
markdown==3.5.2
faiss-cpu
tiktoken 